Replication: generate Replication ID
Once we ensured that both source and target are exists and available for
further processing we need to generate replication ID which will be
used to track this replication process in order to record intermediate
checkpoints and process control.

For Replication ID generation we're using the exact the same approach as
CouchDB has, so our IDs should be compatible. This isn't the exact
requirement for them, but having such feature will allow aiocouchdb to
reuse checkpoints made by CouchDB.

Differences from CouchDB replication implementation:
- Actually, couch_replicator generates ID at the very first step. That
  was correct until there was no need to fetch filter function source
  code, imho, because this process requires to ensure that source peer
  exists and accessible. So couch_replicator eventually may do
  logically the same work twice depending on if filter function
  specified or not. Generation Replication ID on the secod step, when
  both peers are verified makes workflow simpler and a bit solid.
- In case when Replication Checkpoint wasn't find couch_replicator
  downgrades protocol version in order to generate Replication ID
  which could be used by older CouchDB versions. There was no problem
  to add the same functionality for aiocouchdb, but I didn't found
  any reasons to support 3/4/5 years old releases. If you need such
  support - file an issue about.