Replication: find start sequence number
Once we figured what the Replication ID should be we need to search
for Replication Logs on both Source and Peer sides. They may not
be exists - that's ok, Replication will be started from scratch.

If we were lucky and Replication Logs from both peers are available,
we need to compare them.